chore: define `test` util package
