+++ packages/poseidon/package.json
@@ -23,10 +23,7 @@
   "devDependencies": {
     "@ethersproject/bignumber": "^5.7.0",
     "@types/snarkjs": "^0.7.8",
-    "@zk-kit/poseidon-proof": "1.0.0-beta.4",
-    "@zk-kit/utils": "^1.2.0",
-    "ethers": "^6.13.1",
-    "ffjavascript": "^0.3.0",
-    "snarkjs": "^0.7.4"
+    "@zk-kit/test": "workspace:^",
+    "ffjavascript": "^0.3.0"
   }
 }
