+++ packages/test/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@zk-kit/test",
+  "files": ["index.ts"],
+  "main": "index.ts",
+  "private": true,
+  "devDependencies": {
+    "@types/snarkjs": "^0.7.8",
+    "@zk-kit/artifacts": "workspace:^",
+    "@zk-kit/poseidon-proof": "1.0.0-beta.4",
+    "@zk-kit/utils": "^1.2.0",
+    "ethers": "^6.13.1",
+    "snarkjs": "^0.7.4"
+  }
+}
